pyproject.toml
[project]
name = "housemdchatbot-backend"
version = "0.1.0"
description = "A backend for the House MD Chatbot frontend."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "google-genai>=1.41.0",
    "openai>=2.1.0",
    "pinecone>=7.3.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
